+++ docs/troubleshooting/OneNote-Library.ps1
@@ -127,8 +127,8 @@ Function Get-RegAsm-x86-ChildPath {
 
 Function Press-Any-Key
 {
-    if ($Host.Name -notmatch 'ISE') {
-        Write-Host 'Press any key to continue. . .' -NoNewline
+    if (($Host.Name -eq 'ConsoleHost') -and (-not $env:WT_SESSION)) {
+        Write-Host "`nPress any key to continue. . ." -NoNewline
         $Host.UI.RawUI.ReadKey('NoEcho, IncludeKeyDown') | Out-Null
     }
 }