+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## [16.3.0] - 2021-12-05
+
+### Added
+
+- Feature request: configure the notebook used &bull; Issue \#[3] &bull; released from beta
+
+### Changed
+
+- Upgraded OneNote XML schema from 2010 to 2013 version
+
+### Removed
+
+- OneNote 2010 is no longer supported
+
 ## [16.2.1-beta] - 2021-04-07
 
 ### Added
@@ -35,3 +49,4 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 [16.1.0]:https://github.com/atrenton/MyJournal.Notebook/releases/tag/v16.1.0
 [16.2.0]:https://github.com/atrenton/MyJournal.Notebook/releases/tag/v16.2.0
 [16.2.1-beta]:https://github.com/atrenton/MyJournal.Notebook/releases/tag/v16.2.1-beta
+[16.3.0]:https://github.com/atrenton/MyJournal.Notebook/releases/tag/v16.3.0